refactor: move VaildationFailedException to Exceptions folder
